requirements.txt
fastapi==0.115.12
google-genai==1.13.0
httpx==0.28.1
pydantic-settings==2.8.1
sqlalchemy==2.0.40
uvicorn[standard]==0.34.0
